Remove render functions from prelude
