+++ examples/hello/src/main.rs
@@ -28,5 +28,5 @@ fn main() {
     console_error_panic_hook::set_once();
     console_log::init_with_level(log::Level::Debug).unwrap();
 
-    render(|| template! { App() });
+    sycamore::render(|| template! { App() });
 }
