+++ packages/sycamore/src/template.rs
@@ -147,3 +147,22 @@ impl<G: GenericNode> fmt::Debug for Template<G> {
         }
     }
 }
+
+/// Trait for describing how something should be rendered into DOM nodes.
+pub trait IntoTemplate<G: GenericNode> {
+    /// Called during the initial render when creating the DOM nodes. Should return a
+    /// `Vec` of [`GenericNode`]s.
+    fn create(&self) -> Template<G>;
+}
+
+impl<G: GenericNode> IntoTemplate<G> for Template<G> {
+    fn create(&self) -> Template<G> {
+        self.clone()
+    }
+}
+
+impl<T: fmt::Display + ?Sized, G: GenericNode> IntoTemplate<G> for T {
+    fn create(&self) -> Template<G> {
+        Template::new_node(G::text_node(&format!("{}", self)))
+    }
+}
