+++ packages/sycamore/tests/web/keyed.rs
@@ -18,7 +18,7 @@ fn append() {
         }
     });
 
-    render_to(|| node, &test_container());
+    sycamore::render_to(|| node, &test_container());
 
     let p = document().query_selector("ul").unwrap().unwrap();
 
@@ -51,7 +51,7 @@ fn swap_rows() {
         }
     });
 
-    render_to(|| node, &test_container());
+    sycamore::render_to(|| node, &test_container());
 
     let p = document().query_selector("ul").unwrap().unwrap();
     assert_eq!(p.text_content().unwrap(), "123");
@@ -87,7 +87,7 @@ fn delete_row() {
         }
     });
 
-    render_to(|| node, &test_container());
+    sycamore::render_to(|| node, &test_container());
 
     let p = document().query_selector("ul").unwrap().unwrap();
     assert_eq!(p.text_content().unwrap(), "123");
@@ -116,7 +116,7 @@ fn clear() {
         }
     });
 
-    render_to(|| node, &test_container());
+    sycamore::render_to(|| node, &test_container());
 
     let p = document().query_selector("ul").unwrap().unwrap();
     assert_eq!(p.text_content().unwrap(), "123");
@@ -141,7 +141,7 @@ fn insert_front() {
         }
     });
 
-    render_to(|| node, &test_container());
+    sycamore::render_to(|| node, &test_container());
 
     let p = document().query_selector("ul").unwrap().unwrap();
     assert_eq!(p.text_content().unwrap(), "123");
@@ -170,7 +170,7 @@ fn nested_reactivity() {
         }
     });
 
-    render_to(|| node, &test_container());
+    sycamore::render_to(|| node, &test_container());
 
     let p = document().query_selector("ul").unwrap().unwrap();
     assert_eq!(p.text_content().unwrap(), "123");
@@ -203,7 +203,7 @@ fn fragment_template() {
         }
     });
 
-    render_to(|| node, &test_container());
+    sycamore::render_to(|| node, &test_container());
 
     let p = document().query_selector("div").unwrap().unwrap();
 
@@ -250,7 +250,7 @@ fn template_top_level() {
         })
     });
 
-    render_to(|| node, &test_container());
+    sycamore::render_to(|| node, &test_container());
 
     let p = document()
         .query_selector("#test-container")
@@ -296,7 +296,7 @@ fn template_with_other_nodes_at_same_level() {
         }
     });
 
-    render_to(|| node, &test_container());
+    sycamore::render_to(|| node, &test_container());
 
     let elem = document().query_selector("ul").unwrap().unwrap();
